requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
pandas
python-dotenv
pytest
